move resource tracking into own submodule
